Added CRUD routes and tested API
