+++ .env
@@ -1,2 +1,1 @@
 MONGO_URI=mongodb+srv://kishorers86:krish9182740069@cluster1.8ckjz.mongodb.net/?retryWrites=true&w=majority&appName=Cluster1
-PORT=3000
