--- db.js
@@ -1,17 +1,0 @@
-const mongoose = require('mongoose');
-require('dotenv').config();
-
-const connectDB = async () => {
-  try {
-    await mongoose.connect(process.env.MONGO_URI, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    });
-    console.log('✅ MongoDB Connected...');
-  } catch (err) {
-    console.error('❌ MongoDB Connection Error:', err.message);
-    process.exit(1);
-  }
-};
-
-module.exports = connectDB;
