--- docs.bruno/myapicollection/bruno.json
@@ -1,9 +1,0 @@
-{
-  "version": "1",
-  "name": "myapicollection",
-  "type": "collection",
-  "ignore": [
-    "node_modules",
-    ".git"
-  ]
-}
