--- docs.bruno/myapicollection/ping.bru.bru
@@ -1,11 +1,0 @@
-meta {
-  name: ping.bru
-  type: http
-  seq: 2
-}
-
-get {
-  url: http://localhost:3000/ping
-  body: none
-  auth: none
-}
